+++ jsettlers.logic/src/jsettlers/ai/construction/ScoredConstructionPosition.java
@@ -26,9 +26,9 @@ import jsettlers.common.position.ShortPoint2D;
  */
 class ScoredConstructionPosition {
 	ShortPoint2D point;
-	double score;
+	int score;
 
-	public ScoredConstructionPosition(ShortPoint2D point, double score) {
+	public ScoredConstructionPosition(ShortPoint2D point, int score) {
 		this.point = point;
 		this.score = score;
 	}
